refactor(ensjs): move getPrice into actions/public/v1/
Aligns the file layout with sibling v1-only actions (getExpiry, getOwner,
getWrapperData, getWrapperName) which all live under actions/public/v1/ and
are exported only from @ensdomains/ensjs/public/v1.

- move actions/public/getPrice.{ts,test.ts} -> actions/public/v1/
- update relative imports inside moved files
- add getPrice to exports/public/v1.ts barrel
- remove getPrice from top-level exports/public.ts barrel (matches v1 convention)
- update internal callers:
  - actions/wallet/renewNames.test.ts import path
  - query-core/src/actions/public/getPrice.ts re-export source
  - react/src/query/getEnsPrice.ts re-export source
  - JSDoc `@example` imports in registerName.ts, renewNames.ts, getPrice.ts

Breaking: external callers importing `getPrice` from `@ensdomains/ensjs/public`
must now import from `@ensdomains/ensjs/public/v1`.
