+++ packages/ensjs/src/actions/public/v1/getPrice.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest'
-import { publicClient as publicClientL2 } from '../../test/addTestContracts.js'
+import { publicClient as publicClientL2 } from '../../../test/addTestContracts.js'
 import { getPrice } from './getPrice.js'
 
 describe('getPrice', () => {
