+++ packages/ensjs/src/actions/wallet/registerName.ts
@@ -127,7 +127,7 @@ export type RegisterNameErrorType =
  * import { createPublicClient, createWalletClient, http, custom } from 'viem'
  * import { mainnet } from 'viem/chains'
  * import { addEnsContracts } from '@ensdomains/ensjs'
- * import { getPrice } from '@ensdomains/ensjs/public'
+ * import { getPrice } from '@ensdomains/ensjs/public/v1'
  * import { randomSecret } from '@ensdomains/ensjs/utils'
  * import { commitName, registerName } from '@ensdomains/ensjs/wallet'
  *
