+++ packages/ensjs/src/exports/public.ts
@@ -4,7 +4,6 @@ export * from '../actions/public/getAvailable.js'
 export * from '../actions/public/getContentHashRecord.js'
 export * from '../actions/public/getCredentials.js'
 export * from '../actions/public/getName.js'
-export * from '../actions/public/getPrice.js'
 export * from '../actions/public/getRecords.js'
 export * from '../actions/public/getResolver.js'
 export * from '../actions/public/getReverseRecordFromRegistry.js'
